% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef.FormulaRegressor}
\alias{coef.FormulaRegressor}
\title{Extract Model Coefficients}
\usage{
\method{coef}{FormulaRegressor}(x, ...)
}
\arguments{
\item{x}{Object of class inheriting from "FormulaRegressor"}

\item{...}{signature compatibility arguments}
}
\description{
Extract Model Coefficients
}
