% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef.LinearRegression}
\alias{coef.LinearRegression}
\title{Returns estimated coefficients for :class:`LinearRegression` object.}
\usage{
\method{coef}{LinearRegression}(object, ...)
}
\description{
Returns estimated coefficients for :class:`LinearRegression` object.
}
