% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Returns estimated coefficients for :class:`linreg` object.}
\usage{
\method{coef}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{signature compatibility arguments}
}
\description{
Returns estimated coefficients for :class:`linreg` object.
}
