% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{FormulaRegressor-class}
\alias{FormulaRegressor-class}
\alias{FormulaRegressor}
\title{Base FormulaRegressor class.}
\description{
Note: This class should not be used directly. Use derived classes instead.
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula object}

\item{\code{Xy}}{list with X and y arrays}

\item{\code{regressor}}{a :class:`XyRegressor` object}
}}


