% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{get_X_y}
\alias{get_X_y}
\title{Auxiliary function for getting X and y.}
\usage{
get_X_y(formula, data)
}
\arguments{
\item{formula}{an object of class "formula" .}

\item{data}{a data.frame,}
}
\value{
list of X and y arrays.
}
\description{
Auxiliary function for getting X and y.
}
