% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heleramcar4-package.R
\docType{package}
\name{heleramcar4-package}
\alias{heleramcar4}
\alias{heleramcar4-package}
\title{heleramcar4: Implementation of a linear and ridge regression package}
\description{
Implementation of a linear and ridge regression package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/helenallorens/heleramcar4/}
}

}
\author{
\strong{Maintainer}: Helena Llorens \email{hllor282@student.liu.se}

Authors:
\itemize{
  \item Carmen Rodríguez \email{carro162@student.liu.se}
  \item Ramón Amézquita \email{ramam120@student.liu.se}
}

}
\keyword{internal}
