% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{LinearRegressionResults-class}
\alias{LinearRegressionResults-class}
\alias{LinearRegressionResults}
\title{Linear Regression Results}
\description{
Linear Regression Results
}
\section{Fields}{

\describe{
\item{\code{B}}{numeric. Regressions coefficients}

\item{\code{yhat}}{matrix. The fitted values}

\item{\code{residuals}}{matrix. Residuals}

\item{\code{residuals_variance}}{numeric. The residuals variance}

\item{\code{degrees_of_freedom}}{numeric. Degrees of freedom}

\item{\code{B_variance}}{matrix. The variance of the regression coefficients}

\item{\code{t_values}}{numeric. The t-values for each coefficient}

\item{\code{p_values}}{numeric. The p-values for each coefficient}
}}


