% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{LinearRegressor-class}
\alias{LinearRegressor-class}
\alias{LinearRegressor}
\title{Linear Regressor.}
\description{
Linear Regressor.
}
\section{Fields}{

\describe{
\item{\code{coef_}}{numeric.}
}}

\section{Methods}{

\describe{
\item{\code{fit(X, y)}}{Fits Linear Regressor to X and y.}
}}

