% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{plot.LinearRegression}
\alias{plot.LinearRegression}
\title{Returns plot for :class:`LinearRegression` object.}
\usage{
\method{plot}{LinearRegression}(object, ...)
}
\description{
Returns plot for :class:`LinearRegression` object.
}
