% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Returns plot for :class:`linreg` object.}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{signature compatibility arguments}
}
\description{
Returns plot for :class:`linreg` object.
}
