% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred.FormulaRegressor}
\alias{pred.FormulaRegressor}
\title{Returns fitted values for :class:`linreg` object.}
\usage{
\method{pred}{FormulaRegressor}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{signature compatibility arguments}
}
\description{
Returns fitted values for :class:`linreg` object.
}
