% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred.LinearRegression}
\alias{pred.LinearRegression}
\title{Returns fitted values for :class:`LinearRegression` object.}
\usage{
\method{pred}{LinearRegression}(object, ...)
}
\description{
Returns fitted values for :class:`LinearRegression` object.
}
