% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{predict.FormulaRegressor}
\alias{predict.FormulaRegressor}
\title{Predict method for Formula Regressors}
\usage{
\method{predict}{FormulaRegressor}(object, newdata = NULL)
}
\arguments{
\item{object}{Object of class inheriting from "FormulaRegressor"}

\item{newdata}{An optional data frame in which to look for variables with
which to predict. If omitted, the fitted values are used.}
}
\description{
Predict method for Formula Regressors
}
