% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{residuals.FormulaRegressor}
\alias{residuals.FormulaRegressor}
\title{Returns residuals for :class:`FormulaRegressor` object.}
\usage{
\method{residuals}{FormulaRegressor}(x, ...)
}
\arguments{
\item{x}{FormulaRegressor object}

\item{...}{signature compatibility arguments}
}
\description{
Returns residuals for :class:`FormulaRegressor` object.
}
