% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{residuals.LinearRegression}
\alias{residuals.LinearRegression}
\title{Returns residuals for :class:`LinearRegression` object.}
\usage{
\method{residuals}{LinearRegression}(object, ...)
}
\description{
Returns residuals for :class:`LinearRegression` object.
}
