% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Returns residuals for :class:`linreg` object.}
\usage{
\method{residuals}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{signature compatibility arguments}
}
\description{
Returns residuals for :class:`linreg` object.
}
