% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{RidgeRegressor-class}
\alias{RidgeRegressor-class}
\alias{RidgeRegressor}
\title{Ridge Regressor.}
\description{
Ridge Regressor.
}
\section{Fields}{

\describe{
\item{\code{coef_}}{numeric.}

\item{\code{lambda}}{numeric.}
}}

\section{Methods}{

\describe{
\item{\code{fit(X, y)}}{Fits Ridge Regressor to X and y.}
}}

