% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{summary.FormulaRegressor}
\alias{summary.FormulaRegressor}
\title{Returns summary for :class:`FormulaRegressor` object.}
\usage{
\method{summary}{FormulaRegressor}(x, ...)
}
\arguments{
\item{x}{FormulaRegressor object}

\item{...}{signature compatibility arguments}
}
\description{
Returns summary for :class:`FormulaRegressor` object.
}
