% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{summary.LinearRegression}
\alias{summary.LinearRegression}
\title{Returns summary for :class:`LinearRegression` object.}
\usage{
\method{summary}{LinearRegression}(object, ...)
}
\description{
Returns summary for :class:`LinearRegression` object.
}
