% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize Airport Delays}
\usage{
visualize_airport_delays()
}
\description{
This function creates a plot visualizing the mean delay of flights
for different airports, represented by their longitude and latitude.
The data is sourced from the `nycflights13` package, and the function
utilizes `dplyr` for data manipulation and `ggplot2` for visualization.
}
