% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{XyRegressor-class}
\alias{XyRegressor-class}
\alias{XyRegressor}
\title{Base XyRegressor class}
\description{
Note: This class should not be used directly. Use derived classes instead.
}
\section{Fields}{

\describe{
\item{\code{coef_}}{numeric.}
}}

\section{Methods}{

\describe{
\item{\code{predict(X)}}{Predicts input data X}
}}

