% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef.FormulaRegressor}
\alias{coef.FormulaRegressor}
\title{Extract Model Coefficients}
\usage{
\method{coef}{FormulaRegressor}(object, ...)
}
\arguments{
\item{object}{Object of class inheriting from "FormulaRegressor"

#' @export}
}
\description{
Extract Model Coefficients
}
