% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Returns estimated coefficients for :class:`linreg` object.}
\usage{
\method{coef}{linreg}(object, ...)
}
\description{
Returns estimated coefficients for :class:`linreg` object.
}
